pallet_indices::pallet::dispatchables

Function force_transfer

pub fn force_transfer<T: Config>(
    new: <<T as Config>::Lookup as StaticLookup>::Source,
    index: T::AccountIndex,
    freeze: bool,
)

Force an index to an account. This doesn’t require a deposit. If the index is already held, then any deposit is reimbursed to its current owner.

The dispatch origin for this call must be Root.

index: the index to be (re-)assigned.

new: the new owner of the index. This function is a no-op if it is equal to sender.

freeze: if set to true, will freeze the index so it cannot be transferred.

Emits IndexAssigned if successful.

Complexity

O(1).

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::force_transfer.